// src/app/Models/piloto.ts
export interface Piloto {
    // private Long id;

    // private String codigoPiloto;
    // private String nombreCompleto;
    // private char sexo;
    // private LocalTime horasVuelo;

    codigoPiloto: string;
    nombreCompleto: string;
    sexo: string;
    horasVuelo: string;
}
export interface PilotoId {
    id: number;
    codigoPiloto: string;
    nombreCompleto: string;
    sexo: string;
    horasVuelo: string;
}